requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator